backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pandas
openpyxl
reportlab
apscheduler
arabic-reshaper
python-bidi
matplotlib
python-dotenv
requests
redis
qrcode[pil]
slowapi
cryptography
pillow
pydantic-settings
pydantic[email]
email-validator>=2.0.0
alembic
python-json-logger
gunicorn

# ── Observability ─────────────────────
sentry-sdk[fastapi]
prometheus-fastapi-instrumentator

# ── SSO / Push Notifications ─────────
python3-saml
python-ldap
firebase-admin
